check huggingface functional test for nlp
